+++ jsonutil.h
@@ -4,14 +4,15 @@
 #include <string.h>
 #include <ctype.h>
 #include <stdio.h>
+#include <stdint.h>
 #include "math.h"
 
-unsigned int json_popfromantirecursion(JsonContext* context);
+uintptr_t json_popfromantirecursion(JsonContext* context);
 void json_append(const char * data, size_t len, lua_State *L, JsonContext* context, bool isEnd=false);
 void json_bail(lua_State *L, JsonContext* context, const char * err);
 unsigned int table_crc32(const unsigned char* data, int size);
-bool json_addtoantirecursion(unsigned int id, JsonContext* context);
-bool json_existsinantirecursion(unsigned int id, JsonContext* context);
-void json_removefromantirecursion(unsigned int id, JsonContext* context);
+bool json_addtoantirecursion(uintptr_t id, JsonContext* context);
+bool json_existsinantirecursion(uintptr_t id, JsonContext* context);
+void json_removefromantirecursion(uintptr_t id, JsonContext* context);
 void json_pushnullornil(lua_State* L, JsonContext* context);
 bool json_isnull(lua_State* L, JsonContext* context);
